feat: observability layer (dfinity/evm-rpc-canister#370)
Follow-up on dfinity/evm-rpc-canister#364 to add a new observability layer to take care of
logging and metrics. Unfortunately,
[`tower_http::trace`](https://docs.rs/tower-http/latest/tower_http/trace/index.html)
cannot be used in a canister environment because it measures call
latency with `Instant::now`. The proposed layer, while inspired by
[`tower_http::trace`](https://docs.rs/tower-http/latest/tower_http/trace/index.html),
is also simpler because it does not have to deal with streaming
responses.

---------

Co-authored-by: Louis Pahlavi <louis.pahlavi@gmail.com>