build!: remove `ic_cdk` from `evm_rpc_types` dependencies (dfinity/evm-rpc-canister#428)
Remove the `ic_cdk` as a dependency of `evm_rpc_types` to avoid callers
of that library to have a transitive dependency on the `ic_cdk`. This
will help in callers upgrading to `ic_cdk > 0.17` which is incompatible
with previous versions and was the main motivation for dfinity/evm-rpc-canister#416.

**BREAKING CHANGE**: although the interface of the EVM RPC canister did
**not** change, this PR introduces breaking changes for Rust clients of
the following libraries:
1. `canhttp`: the `IcError` type now uses `ic_error_types::RejectCode`
instead of `ic_cdk::api::call::RejectionCode`
2. `evm_rpc_types`: the variant `HttpOutcallError::IcError` now uses a
local copy of `ic_cdk::api::call::RejectionCode`, renamed to
`LegacyRejectionCode`. The Candid serialization/deserialization of that
type should remain identical.