feat: Select supported providers based on successful responses (dfinity/evm-rpc-canister#435)
Dynamically select supported JSON-RPC providers according to the number
of successful JSON-RPC responses in the last 20 minutes, where providers
with a higher number will be prioritized over ones with lower numbers.

This is a heuristic that aims at selecting only currently functioning
providers. The idea is that since JSON-RPC errors cannot be relied upon,
as they are not specified by the Ethereum JSON-RPC API and so different
EVM clients or JSON-RPC providers do use different errors, we only count
instead the number of successful responses. Obviously, this is still not
bulletproof, since it won't help in case the provider returns a
successful JSON-RPC response with incorrect data.

Looking back at previous incidents with Ethereum JSON-RPC providers for
the EVM RPC canister or the ckETH/ckERC20 minter, we can see that having
this heuristic would have helped in most cases, excepted when a provider
returns successful responses with incorrect data (1 out of 4 incidents):

1. :relieved: `Cloudflare` always return an error (proposal
[136701](https://dashboard.internetcomputer.org/proposal/136701))
2. :relieved: `LlamaNodes` is down (proposal
[132474](https://dashboard.internetcomputer.org/proposal/132474))
3. :relieved: `Ankr` dropped IPv6 connectivity
([132415](https://dashboard.internetcomputer.org/proposal/132415)
4. :scream:`Cloudflare` returns successful responses with wrong results
(proposal
[128365](https://dashboard.internetcomputer.org/proposal/128365))

This is the last step towards dynamically ranking supported providers
according to the number of their successful responses, which is done in
the following 3 PRs:

1. dfinity/evm-rpc-canister#436
5. dfinity/evm-rpc-canister#434 
6. dfinity/evm-rpc-canister#435 (this one)